Rename game_state.rs to more appropriate clash_game.rs
